requirements.txt
requests==2.31.0
python-telegram-bot==13.13
deep-translator==1.11.4
termcolor==2.3.0
python-dotenv==1.0.0